Moved capitalize from single file to mod file
